sys/src/9k/port: use uintptr for the qmalloc allocator family and internals (thanks Geoff Collyer)

+++ sys/src/9k/port/portfns.h
@@ -306,7 +306,7 @@ int		setlabel(Label*);
 void		setregisters(Ureg*, char*, char*, int);
 char*		skipslash(char*);
 void		sleep(Rendez*, int (*)(void*), void*);
-void*		smalloc(ulong);
+void*		smalloc(uintptr);
 char*		srvname(Chan*);
 void		syscallfmt(int, va_list list);
 void		sysretfmt(int, va_list, Ar0*, uvlong, uvlong);
